+++ MerfToSDPForJava.sln
@@ -7,9 +7,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MerfToSDPForJava", "src\Mer
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{4629663B-032C-4FF6-8EB9-C821E9FD8952}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Content.Build.MerfToSDPForJava.DataContracts", "src\Microsoft.Content.Build.MerfToSDPForJava.DataContracts\Microsoft.Content.Build.MerfToSDPForJava.DataContracts.csproj", "{259ADCB6-608A-4B58-97D7-74BF5798D7C3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts", "src\Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts\Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts.csproj", "{92EE20FF-2088-429A-9A27-314BFDDC3E4F}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Content.Build.MerfToSDPForJava.DataContracts", "src\Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts\Microsoft.Content.Build.MerfToSDPForJava.DataContracts.csproj", "{92EE20FF-2088-429A-9A27-314BFDDC3E4F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Content.Build.MerfToSDPForJava.Constants", "src\Microsoft.Content.Build.MerfToSDPForJava.Constants\Microsoft.Content.Build.MerfToSDPForJava.Constants.csproj", "{FF02FCF1-1C69-450E-B531-1A52F3042841}"
 EndProject
@@ -23,10 +21,6 @@ Global
 		{57657B28-4E7F-4CD3-BD06-A0DBB2D6B7A0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{57657B28-4E7F-4CD3-BD06-A0DBB2D6B7A0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{57657B28-4E7F-4CD3-BD06-A0DBB2D6B7A0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{259ADCB6-608A-4B58-97D7-74BF5798D7C3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{259ADCB6-608A-4B58-97D7-74BF5798D7C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{259ADCB6-608A-4B58-97D7-74BF5798D7C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{259ADCB6-608A-4B58-97D7-74BF5798D7C3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{92EE20FF-2088-429A-9A27-314BFDDC3E4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{92EE20FF-2088-429A-9A27-314BFDDC3E4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{92EE20FF-2088-429A-9A27-314BFDDC3E4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -41,7 +35,6 @@ Global
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{57657B28-4E7F-4CD3-BD06-A0DBB2D6B7A0} = {4629663B-032C-4FF6-8EB9-C821E9FD8952}
-		{259ADCB6-608A-4B58-97D7-74BF5798D7C3} = {4629663B-032C-4FF6-8EB9-C821E9FD8952}
 		{92EE20FF-2088-429A-9A27-314BFDDC3E4F} = {4629663B-032C-4FF6-8EB9-C821E9FD8952}
 		{FF02FCF1-1C69-450E-B531-1A52F3042841} = {4629663B-032C-4FF6-8EB9-C821E9FD8952}
 	EndGlobalSection
