--- src/Microsoft.Content.Build.MerfToSDPForJava.DataContracts/Hierarchy/HierarchyType.cs
@@ -1,11 +1,0 @@
-﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts
-{
-    public enum HierarchyType
-    {
-        Namespace,
-        Class,
-        Struct,
-        Enum,
-        Interface,
-    }
-}
