--- src/Microsoft.Content.Build.MerfToSDPForJava.DataContracts/Microsoft.Content.Build.MerfToSDPForJava.DataContracts.csproj
@@ -1,11 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="YamlDotNet" Version="8.1.2" />
-  </ItemGroup>
-
-</Project>
