--- src/Microsoft.Content.Build.MerfToSDPForJava.DataContracts/ViewModel/MemberType.cs
@@ -1,31 +1,0 @@
-﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts
-{
-    using System;
-
-    [Serializable]
-    public enum MemberType
-    {
-        Namespace,
-        Class,
-        Struct,
-        Enum,
-        Field,
-        Method,//MemberFunction
-        //Friend,
-        Interface,
-        Constructor,
-    }
-
-    [Flags]
-    public enum AccessLevel
-    {
-        None = 0,
-        Private = 1,
-        Public = 2,
-        Protected = 4,
-        Static = 8,
-        Virtual = 16,
-        Package = 32,
-        NotAccessible = Private | Package,
-    }
-}
