--- src/Microsoft.Content.Build.MerfToSDPForJava.DataContracts/ViewModel/SpecializedType.cs
@@ -1,15 +1,0 @@
-﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts
-{
-    using System;
-    using YamlDotNet.Serialization;
-
-    [Serializable]
-    public class SpecializedType
-    {
-        [YamlMember(Alias = "type")]
-        public string Type { get; set; }
-
-        [YamlMember(Alias = "name")]
-        public string SpecializedFullName { get; set; }
-    }
-}
