--- src/Microsoft.Content.Build.MerfToSDPForJava.DataContracts/ViewModel/TocItemYaml.cs
@@ -1,28 +1,0 @@
-﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts
-{
-    using System;
-
-    using YamlDotNet.Serialization;
-
-    [Serializable]
-    public class TocItemYaml
-    {
-        [YamlMember(Alias = "uid")]
-        public string Uid { get; set; }
-
-        [YamlMember(Alias = "name")]
-        public string Name { get; set; }
-
-        [YamlMember(Alias = "href")]
-        public string Href { get; set; }
-
-        [YamlMember(Alias = "originalHref")]
-        public string OriginalHref { get; set; }
-
-        [YamlMember(Alias = "homepage")]
-        public string Homepage { get; set; }
-
-        [YamlMember(Alias = "items")]
-        public TocYaml Items { get; set; }
-    }
-}
