--- src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/ItemType.cs
@@ -1,25 +1,0 @@
-﻿namespace Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts
-{
-    using System;
-    public enum ItemType
-    {
-        Default,
-        Toc,
-        Assembly,
-        Namespace,
-        Class,
-        Interface,
-        Struct,
-        Delegate,
-        Enum,
-        Field,
-        Property,
-        Event,
-        Constructor,
-        Method,
-        Operator,
-        Container,
-        AttachedEvent,
-        AttachedProperty
-    }
-}
