+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Merf/ApiParameter.cs
@@ -1,0 +1,18 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.Merf
+{
+    using System;
+    using YamlDotNet.Serialization;
+
+    [Serializable]
+    public class ApiParameter
+    {
+        [YamlMember(Alias = "id")]
+        public string Name { get; set; }
+
+        [YamlMember(Alias = "type")]
+        public string Type { get; set; }
+
+        [YamlMember(Alias = "description")]
+        public string Description { get; set; }
+    }
+}
