+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Merf/CrefInfo.cs
@@ -1,0 +1,15 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.Merf
+{
+    using System;
+    using YamlDotNet.Serialization;
+
+    [Serializable]
+    public class CrefInfo
+    {
+        [YamlMember(Alias = "type")]
+        public string Type { get; set; }
+
+        [YamlMember(Alias = "description")]
+        public string Description { get; set; }
+    }
+}
