+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Merf/MemberType.cs
@@ -1,0 +1,30 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.Merf
+{
+    using System;
+
+    [Serializable]
+    public enum MemberType
+    {
+        Namespace,
+        Class,
+        Struct,
+        Enum,
+        Field,
+        Method,//MemberFunction
+        Interface,
+        Constructor,
+    }
+
+    [Flags]
+    public enum AccessLevel
+    {
+        None = 0,
+        Private = 1,
+        Public = 2,
+        Protected = 4,
+        Static = 8,
+        Virtual = 16,
+        Package = 32,
+        NotAccessible = Private | Package,
+    }
+}
