+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Merf/SpecViewModel.cs
@@ -1,0 +1,25 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.Merf
+{
+    using System;
+
+    using YamlDotNet.Serialization;
+
+    [Serializable]
+    public class SpecViewModel
+    {
+        [YamlMember(Alias = "uid")]
+        public string Uid { get; set; }
+
+        [YamlMember(Alias = "name")]
+        public string Name { get; set; }
+
+        [YamlMember(Alias = "fullName")]
+        public string FullName { get; set; }
+
+        [YamlMember(Alias = "isExternal")]
+        public bool IsExternal { get; set; }
+
+        [YamlMember(Alias = "href")]
+        public string Href { get; set; }
+    }
+}
