+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Merf/TocYaml.cs
@@ -1,0 +1,13 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.Merf
+{
+    using System;
+    using System.Collections.Generic;
+
+    [Serializable]
+    public class TocYaml
+        : List<TocItemYaml>
+    {
+        public TocYaml(IEnumerable<TocItemYaml> items) : base(items) { }
+        public TocYaml() : base() { }
+    }
+}
