+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Microsoft.Content.Build.MerfToSDPForJava.DataContracts.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    <PackageReference Include="YamlDotNet" Version="8.1.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Microsoft.Content.Build.MerfToSDPForJava.Constants\Microsoft.Content.Build.MerfToSDPForJava.Constants.csproj" />
+  </ItemGroup>
+
+</Project>
