--- src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts.csproj
@@ -1,16 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Microsoft.Content.Build.MerfToSDPForJava.Constants\Microsoft.Content.Build.MerfToSDPForJava.Constants.csproj" />
-    <ProjectReference Include="..\Microsoft.Content.Build.MerfToSDPForJava.DataContracts\Microsoft.Content.Build.MerfToSDPForJava.DataContracts.csproj" />
-  </ItemGroup>
-
-</Project>
