+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/SDP/Shared/EnumField.cs
@@ -1,0 +1,28 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.SDP
+{
+    using Newtonsoft.Json;
+    using YamlDotNet.Serialization;
+    public class EnumField
+    {
+        [JsonProperty("fullName")]
+        [YamlMember(Alias = "fullName")]
+        public string FullName { get; set; }
+
+        [JsonProperty("name")]
+        [YamlMember(Alias = "name")]
+        public string Name { get; set; }
+
+        [JsonProperty("nameWithType")]
+        [YamlMember(Alias = "nameWithType")]
+        public string NameWithType { get; set; }
+
+        [JsonProperty("summary")]
+        [YamlMember(Alias = "summary")]
+        public string Summary { get; set; }
+
+        [JsonProperty("uid")]
+        [YamlMember(Alias = "uid")]
+        public string Uid { get; set; }
+
+    }
+}
