+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/SDP/Shared/Member.cs
@@ -1,0 +1,12 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.SDP
+{
+    using Newtonsoft.Json;
+    using System.Collections.Generic;
+    using YamlDotNet.Serialization;
+    public class Member:EnumMethod
+    {
+         [JsonProperty("typeParameters")]
+        [YamlMember(Alias = "typeParameters")]
+        public IEnumerable<TypeParameter> TypeParameters { get; set; }
+    }
+}
