+++ src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/SDP/Shared/Return.cs
@@ -1,0 +1,15 @@
+﻿namespace Microsoft.Content.Build.MerfToSDPForJava.DataContracts.SDP
+{
+    using Newtonsoft.Json;
+    using YamlDotNet.Serialization;
+    public class Return
+    {
+        [JsonProperty("description")]
+        [YamlMember(Alias = "description")]
+        public string Description { get; set; }
+
+        [JsonProperty("type")]
+        [YamlMember(Alias = "type")]
+        public string Type { get; set; }
+    }
+}
