--- src/Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts/Shared/TypeParameter.cs
@@ -1,15 +1,0 @@
-﻿namespace Microsoft.Content.Build.MerfToSDPForJava.SDP.DataContracts
-{
-    using Newtonsoft.Json;
-    using YamlDotNet.Serialization;
-    public class TypeParameter
-    {
-        [JsonProperty("description")]
-        [YamlMember(Alias = "description")]
-        public string Description { get; set; }
-
-        [JsonProperty("name")]
-        [YamlMember(Alias = "name")]
-        public string Name { get; set; }
-    }
-}
